src/functions/requirements.txt
azure-functions
azure-identity
azure-storage-file-datalake
azure-keyvault-secrets
pydantic>=2.0
pyarrow
pandas
